feat(cmbLabel): Add a mergeRangeLabelsIfSame  option
In 3afcce9 commit, I changed the behavior of the combined label for range slider. However, this
broke previous implementation, so this commit add an option to configure this feature. The default
behavior is back as previously: If the slider values are the same, the combined label is 'X - X'

Related to #245